Add a test case of how much clang is able to inline/optimize out
The question is basically whether e.g. array[i] = func(array, i)
in do_all_two can be turned into just a plain array[i] = array[i] + 1,
and the answer is definitely 'no', at least for now.

The problem is that constructing the GC frame and such is pretty
significant; it's not something that can just be optimized out.

I'm not sure the best way to make this thing optimizeable away.
One way is of course to simply not generate GC frames when
we can at all avoid it. But it is very strange.