I think I figured out my issue..
So I was thinking that the only valid solutions to Part 2 would
be solutions that would also (sort of) be Part 1 solutions.

THis doesn't actually seem to be the case. Instead, *any*
combination of buttons is allowed. This is in some ways
better and in some ways worse.

The way it's better is that we don't have to DFS over multiple
solutions. The way it's worse is that we have to do a much
deeper search. :(